-- feature  - watcher event not trigger by api call instead of server started  - stream logging pod log and record into db when server started  - join logging stream when new pod applied  - make all namespaces event watcher and pod logging only once

+++ resource/immigration.sql
@@ -17,7 +17,7 @@ create table if not exists `user`
 create table if not exists logs
 (
     id          int auto_increment,
-    hash_code   int         default 0     not null,
+    hash_code   bigint         default 0     not null,
     app         varchar(20) default ''    not null,
     pod         varchar(36) default ''    not null,
     content     text                      not null,
